requirements.txt
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.4.2
pyyaml==6.0.1
python-levenshtein==0.25.1
rapidfuzz==3.6.1
matplotlib==3.8.4
flake8==7.0.0
